requirements.txt
pytorch-lightning>=2.0.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=5.4.1
tqdm>=4.61.0
scikit-learn>=0.24.0
monai~=1.1.0
albumentations>=1.0.0
tensorboard>=2.5.0
addict
timm
openpyxl